SCRUM-127 create API POST payment, file PaypalConfig

+++ src/main/java/com/TrungTinhBackend/codearena_backend/Entity/User.java
@@ -44,7 +44,7 @@ public class User implements UserDetails {
 
     private Long point;
 
-    private Long coin;
+    private Double coin;
 
     private RankEnum rankEnum;
