+++ src/main/java/com/TrungTinhBackend/codearena_backend/Request/APIRequestUserUpdate.java
@@ -27,7 +27,7 @@ public class APIRequestUserUpdate {
 
     private Long point;
 
-    private Long coin;
+    private Double coin;
 
     private RankEnum rankEnum;
